refractor processing functions from controller to applications
